% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi2.R
\name{multi2}
\alias{multi2}
\title{Multiplica dos numeros}
\usage{
multi2(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\description{
esta funcion multiplica dos numeros
}
\examples{
multi2(a=2,b=4)
multi2(a=3,b=5)
}

