% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma2.R
\name{suma2}
\alias{suma2}
\title{Suma dos numeros}
\usage{
suma2(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\description{
Esta función suma dos numeros
}
\examples{
suma2(a=3,b=4)
suma2(a=5,b=6)
}

